requirements.txt
# Dependencies untuk Website Deteksi Penyakit Bawang Merah
# Install dengan: pip install -r requirements.txt

# Core web framework
Flask==2.3.3
Flask-CORS==4.0.0

# Production WSGI server
gunicorn==21.2.0

# Image processing
Pillow==10.0.0

# Utilities
requests==2.31.0

# Optional: AI/ML (untuk development model)
# tensorflow==2.13.0
# scikit-learn==1.3.0
# numpy==1.24.3